rename ResetTracker.py to reset_tracker.py
(and update all calls to it)
