+++ hints/gui/reset_tracker.py
